stage #3/java/module #1. GIT & Build Tools/build tools/task2/utils/src/test/java/by/tkachuk/utils/StringUtilsTest.java
package by.tkachuk.utils;


import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;


public class StringUtilsTest {
    @Test
    public void isPositiveNumberTest() {
        String test = "12";
        Assertions.assertTrue(StringUtils.isPositiveNumber(test));
    }

    @Test
    public void isNotPositiveNumberTest() {
        String test = "-12";
        Assertions.assertFalse(StringUtils.isPositiveNumber(test));
    }

    @Test()
    public void isPositiveNumberTestWithNullParam() {
        Assertions.assertThrows(IllegalArgumentException.class,
                () -> StringUtils.isPositiveNumber(null));
    }


}
